[REF] Tidy up sources
