+++ spank/nccl-inspector-preconf/src/snccliprecon_config.h
@@ -1,11 +1,10 @@
 #ifndef SNCCLIPRECON_CONFIG_H
 #define SNCCLIPRECON_CONFIG_H
 
-#include <stdbool.h>
-#include <slurm/spank.h>
-#include <stddef.h>
+#include "snccliprecon.h"
 
-#define SNCCLIPRECON_PATH_MAX 4096
+#include <stdbool.h>
+#include <stddef.h>
 
 typedef struct snccliprecon_config {
   bool enabled;
@@ -23,5 +22,22 @@ void snccliprecon_config_reset(void);
 void snccliprecon_config_parse_args(spank_t spank, int argc, char **argv);
 
 spank_err_t snccliprecon_config_parse_option(const char *name, const char *value);
+
+spank_err_t snccliprecon_args_register(spank_t spank);
+
+int snccliprecon_config_render_job_dump_dir(
+  const snccliprecon_config_t *config,
+  const char *job_id,
+  char *output,
+  size_t output_size
+);
+
+int snccliprecon_config_render_step_dump_dir(
+  const snccliprecon_config_t *config,
+  const char *job_id,
+  const char *step_id,
+  char *output,
+  size_t output_size
+);
 
 #endif
