+++ spank/nccl-inspector-preconf/src/snccliprecon_worker_lock.h
@@ -1,21 +1,8 @@
-#ifndef SNCCLIPRECON_FILE_OPS_H
-#define SNCCLIPRECON_FILE_OPS_H
+#ifndef SNCCLIPRECON_WORKER_LOCK_H
+#define SNCCLIPRECON_WORKER_LOCK_H
 
-#include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
-
-typedef int (*snccliprecon_locked_fn_t)(void *arg);
-
-int snccliprecon_ensure_mode(const char *path);
-
-bool snccliprecon_dir_exists(const char *path);
-
-int snccliprecon_mkdir_all(const char *path);
-
-int snccliprecon_ensure_dir(const char *path);
-
-int snccliprecon_with_lock(const char *path, snccliprecon_locked_fn_t fn, void *arg);
 
 int snccliprecon_get_hostname(char *hostname, size_t hostname_size);
