--- spank/nccl-inspector-preconf/src/spank_constants.h
@@ -1,13 +1,0 @@
-#ifndef SNCCLIPRECON_CONSTANTS_H
-#define SNCCLIPRECON_CONSTANTS_H
-
-#define SNCCLIPRECON_DEFAULT_MODE 0777
-#define SNCCLIPRECON_TMP_DIR_BASE "/tmp/nccl_inspector_preconf"
-
-#define SNCCLIPRECON_USER_INIT_OP "user-init"
-#define SNCCLIPRECON_TASK_INIT_PRIVILEGED_OP "task-init-privileged"
-#define SNCCLIPRECON_TASK_EXIT_OP "task-exit"
-
-#define SNCCLIPRECON_LOG_DIR_SET_BY_PLUGIN "SNCCLIPRECON_LOG_DIR_SET_BY_PLUGIN"
-
-#endif
